requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
yfinance>=0.2.0
ta>=0.10.2
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
